+++ CHANGELOG.md
@@ -24,3 +24,21 @@ Added the backdrop feature:
 Other changes:
 - Removed the README from the example app (pub will display the code on the website now)
 - Specified Dart as the language in the README code snippets
+
+## [0.2.1] - [TODO: Add Release Date]
+
+Added callbacks to the `SlidingUpPanel`
+- `onPanelSlide`
+- `onPanelOpened`
+- `onPanelCollapsed`
+
+Added methods to the `PanelController`
+- `setPanelPosition`
+- `animatePanelToPosition`
+- `getPanelPosition`
+- `isPanelAnimating`
+- `isPanelOpen`
+- `isPanelCollapsed`
+- `isPanelShown`
+
+Updated documentation to reflect changes